fixed gear change bug
From Drive gear, when the car is moving, the gear cannot be changed to Park or Reverse.

In Reverse gear, when the car is moving, the gear cannot be changed to Drive or Park

Also fixed the gear change to Drive bug at the first gear change.

Changed the RPM intervals in shift change for TEST